Idle if Nothing to Do
Basically implemented thread blocking/sleeping. Put a thread to sleep if
it's waiting for a process or keyboard event. Run an idle thread if
there's no other thread that can be run. This thread halts the CPU until
a tick happens. If a thread has been awoken, then disable the idle
thread and run the awakened thread.