Fix Zig Errors in System Calls
When using cat with an invalid file, an unexpected error message would
be printed by cat. In the kernel the Zig error going back from the
system call was correct, but the user program didn't seem to understand

The reason for this is Zig errors can only safely be used within the
same compilation. To get around that system calls use ValueOrError type
to translate the kernel Zig errors to an errno-like ABI-stable enum to
pass across the system call boundary.  In user space, the same type
tries to translate the enum back to Zig errors. If the kernel error type
is unknown to the user program it gets the utils.Error.Unknown error.

Converted the exec, open, read, and close syscalls to use this. Still
need to convert the next_dir_entry system call.