Set VBE/VESA Modes to Get High Resolution Graphics
Using make vbe=true, it's possible for Georgios to set the VBE mode on
it's own without GRUB doing it for us like when using the
multiboot_vbe=true build option. It uses libx86emu to emulate a CPU to
invoke BIOS interrupts to all this.

This is still buggy though. The emulated CPU (double emulated CPU in
QEMU) needs to be slowed when setting the mode or else it doesn't work,
at least in QEMU. With this, it works in QEMU, but won't update the
screen with my test graphics until the emulator window is messed with
somehow, like putting it to full screen. My test laptop panics when
running the mode setting code, but this should be easy to fix once I
find the overflow causing it. Need to test bochs.

Once these bugs are worked out, mode picking is more flexible, and there
is a VBE console I'm satisfied with, vbe build option will become
true by default.