Guard Pages for Kernel Mode Stacks
To protect against kernel stack overflow, add an intentionally missing
page for a kernel mode stack to hit if it gets too big. This causes a
page fault, but for right now this just leads to a triple fault because
it's already on the stack for handling interrupts and the stack is
unusable. This is a bit better than doing something random like
corrupting the copy of page directory of the running process like it is
right now. Still, it would be very nice to at least have some sort of
indication of what happened or even better a way to see the backtrace of
it in gdb.