Fix Corruption in I/O Block Cache
This was caused by the explict free_block implemention of the block
cache freeing the block data with the real block store but leaving the
block otherwise unchanged in the cache. This only gets called once in
gpt code, but the memory for the block data gets reused. This casues
subsequent blocks using the memory to have a chance to be corrupted,
which was subtle and didn't really show up unless there was a lot of
disk reads or the cache size was reduced so the corrupted memory of the
cache had a larger chance to be used.

Along the way I added things to debug this. This includes the
direct_disk build option to skip the block cache and test-file and
check-test-file to test reading a large (1 MiB) file.

Also added a check to ATA to make sure a sector address wasn't too
large. This was discovered after Ext2 was trying to deal with corrected
filesystem data from the cache. It looks like an invalid address casues
the drive to be unusable afterwards, at least on QEMU.