UnitTestAllocator Now Uses GeneralPurposeAllocator
This is done through utils.TestAlloc. TestAlloc will probably replace
UnitTestAllocator when kernel.memory.Allocator is replaced with
std.mem.Allocator.

Also added StringWriter, which I was going to use for making stack
traces for leaked memory, but GeneralPurposeAllocator already does this.
It might be useful for something in the future though.