Adding Dispatching and mkdir/touch/unlink
Dispatching is the planned future method for IPC and generally how
programs will talk to the OS. The lowest layer of this are ports for
passing bytes in messages called `Dispatch` to and from other ports.
Interfaces in OMG IDL are used to generate abstract structs in Zig with
helper code to pass parameters of the abstract methods using a port.
This can then interact directly with kernel code. This should be
somewhat easier then implementing system calls using the current method.

Later this should also be able to be used to communicate with other
threads and processes using queued messages. Accessing interfaces will
be done through the file system where objects (could be files or
directories) will present multiple possible interfaces. Current
influences for all this are Mach ports and to a lesser extend Fuchsia
interfaces and Plan 9 Styx.

Used this very crude initial version to implement mkdir, touch, and
unlink commands using the added `Directory` interface.

Also:
- Removed the disk2.img ATA write test that required qemu-img.
- Trying to create a file/directory when the name is already used will
  result in an error.